requirements.txt
google-generativeai
langchain-google-genai
requests
beautifulsoup4
crewai
streamlit
streamlit_option_menu
pdfplumber